requirements.txt
opencv-python
numpy
pyaudio
wave
picamera2
ffmpeg-python
telebot
python-dotenv